app/requirements.txt
gunicorn
asgiref==3.7.2
confluent-kafka==2.3.0
Django==4.2.10
djangorestframework==3.14.0
freezegun==1.4.0
pika==1.3.2
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
